+++ devolo_home_control_api/homecontrol.py
@@ -1,6 +1,8 @@
 import threading
+from typing import Optional
 
 import requests
+from zeroconf import Zeroconf
 
 from . import __version__
 from .backend.mprm import Mprm
@@ -28,16 +30,18 @@ class HomeControl(Mprm):
     Unit, your devices and their properties.
 
     :param gateway_id: Gateway ID (aka serial number), typically found on the label of the device
+    :param zeroconf_instance: Zeroconf instance to be potentially reused
     :param url: URL of the mPRM (typically leave it at default)
     """
 
-    def __init__(self, gateway_id: str, url: str = "https://homecontrol.mydevolo.com"):
+    def __init__(self, gateway_id: str, zeroconf_instance: Optional[Zeroconf] = None,
+                 url: str = "https://homecontrol.mydevolo.com"):
         self._session = requests.Session()
         self._session.headers.update({"User-Agent": f"devolo_home_control_api/{__version__}"})
-        self._session.url = url
+        self._session.url = url.rstrip("/")
 
         self.gateway = Gateway(gateway_id)
-        super().__init__()
+        super().__init__(zeroconf_instance)
 
         self.gateway.zones = self.get_all_zones()
 
@@ -144,7 +148,8 @@ class HomeControl(Mprm):
             BinarySwitchProperty(session=self._session,
                                  gateway=self.gateway,
                                  element_uid=uid_info['UID'],
-                                 state=bool(uid_info['properties']['state']))
+                                 state=bool(uid_info['properties']['state']),
+                                 enabled=uid_info['properties']['guiEnabled'])
 
     def _general_device(self, uid_info: dict):
         """ Process general device setting (gds) properties. """
@@ -211,6 +216,7 @@ class HomeControl(Mprm):
                     "devolo.SirenMultiLevelSensor": self._multi_level_sensor,
                     "devolo.SirenMultiLevelSwitch": self._multi_level_switch,
                     "devolo.ShutterMovementFI": self._binary_sensor,
+                    "devolo.ValveTemperatureSensor": self._multi_level_sensor,
                     "devolo.VoltageMultiLevelSensor": self._multi_level_sensor,
                     "devolo.WarningBinaryFI": self._binary_sensor,
                     "acs.hdm": self._automatic_calibration,