+++ devolo_home_control_api/properties/binary_switch_property.py
@@ -1,9 +1,10 @@
 from datetime import datetime
+from typing import Any
 
 from requests import Session
 
 from ..devices.gateway import Gateway
-from ..exceptions.device import WrongElementError
+from ..exceptions.device import SwitchingProtected, WrongElementError
 from .property import Property
 
 
@@ -14,16 +15,18 @@ class BinarySwitchProperty(Property):
     :param gateway: Instance of a Gateway object
     :param session: Instance of a requests.Session object
     :param element_uid: Element UID, something like devolo.BinarySwitch:hdm:ZWave:CBC56091/24#2
-    :param state: State the switch has at time of creating this instance
+    :key enabled: State of the remote protection setting
+    :key state: State the switch has at time of creating this instance
     """
 
-    def __init__(self, gateway: Gateway, session: Session, element_uid: str, state: bool):
+    def __init__(self, gateway: Gateway, session: Session, element_uid: str, **kwargs: Any):
         if not element_uid.startswith("devolo.BinarySwitch:"):
             raise WrongElementError(f"{element_uid} is not a Binary Switch.")
 
         super().__init__(gateway=gateway, session=session, element_uid=element_uid)
 
-        self._state = state
+        self._state = kwargs.get("state", False)
+        self.enabled = kwargs.get("enabled", False)
 
 
     @property
@@ -44,6 +47,9 @@ class BinarySwitchProperty(Property):
 
         :param state: True if switching on, False if switching off
         """
+        if not self.enabled:
+            raise SwitchingProtected("This device is protected against remote switching.")
+
         data = {"method": "FIM/invokeOperation",
                 "params": [self.element_uid, "turnOn" if state else "turnOff", []]}
         response = self.post(data)