+++ tests/mocks/mock_dnsrecord.py
@@ -1,4 +1,0 @@
-class MockDNSRecord:
-    def __init__(self):
-        self.address = ""
-        self.key = "devolo-homecontrol"