+++ tests/mocks/mock_metering_plug.py
@@ -23,8 +23,8 @@ def metering_plug(device_uid: str) -> Zwave:
     with file.open("r") as fh:
         test_data = json.load(fh)
 
-    device = Zwave(**test_data.get("devices").get("mains").get("properties"))
-    gateway = MockGateway(test_data.get("gateway").get("id"))
+    device = Zwave(**test_data['devices']['mains']['properties'])
+    gateway = MockGateway(test_data['gateway']['id'])
     session = requests.Session()
 
     device.binary_switch_property = {}
@@ -36,53 +36,42 @@ def metering_plug(device_uid: str) -> Zwave:
         BinarySwitchProperty(gateway=gateway,
                              session=session,
                              element_uid=f"devolo.BinarySwitch:{device_uid}",
-                             state=test_data.get("devices").get("mains").get("properties").get("state"))
-    device.consumption_property[f'devolo.Meter:{device_uid}'] = ConsumptionProperty(gateway=gateway,
-                                                                                    session=session,
-                                                                                    element_uid=f"devolo.Meter:{device_uid}",
-                                                                                    current=test_data.get("devices").get(
-                                                                                        "mains").get("properties").get(
-                                                                                        "current_consumption"),
-                                                                                    total=test_data.get("devices").get(
-                                                                                        "mains").get("properties").get(
-                                                                                        "total_consumption"),
-                                                                                    total_since=test_data.get("devices").get(
-                                                                                        "mains").get("properties").get(
-                                                                                        "sinceTime"))
+                             state=test_data['devices']['mains']['properties']['state'],
+                             enabled=test_data['devices']['mains']['properties']['guiEnabled'])
+    device.consumption_property[f'devolo.Meter:{device_uid}'] = \
+        ConsumptionProperty(gateway=gateway,
+                            session=session,
+                            element_uid=f"devolo.Meter:{device_uid}",
+                            current=test_data['devices']['mains']['properties']['current_consumption'],
+                            total=test_data['devices']['mains']['properties']['total_consumption'],
+                            total_since=test_data['devices']['mains']['properties']['sinceTime'])
     device.multi_level_sensor_property[f'devolo.VoltageMultiLevelSensor:{device_uid}'] = \
         MultiLevelSensorProperty(gateway=gateway,
                                  session=session,
                                  element_uid=f"devolo.VoltageMultiLevelSensor:{device_uid}",
-                                 current=test_data.get("devices").get("mains").get("properties").get("voltage"))
-    device.settings_property["param_changed"] = SettingsProperty(gateway=gateway,
-                                                                 session=session,
-                                                                 element_uid=f"cps.{device_uid}")
-    device.settings_property["general_device_settings"] = SettingsProperty(gateway=gateway,
-                                                                           session=session,
-                                                                           element_uid=f"gds.{device_uid}",
-                                                                           events_enabled=test_data.get("devices").get(
-                                                                               "mains").get("properties").get(
-                                                                               "events_enabled"),
-                                                                           icon=test_data.get("devices").get(
-                                                                               "mains").get("properties").get(
-                                                                               "icon"),
-                                                                           name=test_data.get("devices").get(
-                                                                               "mains").get("properties").get(
-                                                                               "itemName"),
-                                                                           zone_id=test_data.get("devices").get(
-                                                                               "mains").get("properties").get(
-                                                                               "zoneId"))
-    device.settings_property["led"] = SettingsProperty(gateway=gateway,
-                                                       session=session,
-                                                       element_uid=f"lis.{device_uid}",
-                                                       led_setting=test_data.get("devices").get("mains").get(
-                                                           "properties").get("led_setting"))
-    device.settings_property["protection"] = SettingsProperty(gateway=gateway,
-                                                              session=session,
-                                                              element_uid=f"ps.{device_uid}",
-                                                              local_switching=test_data.get("devices").get("mains").get(
-                                                                  "properties").get("local_switch"),
-                                                              remote_switching=test_data.get("devices").get("mains").get(
-                                                                  "properties").get("remote_switch"))
+                                 current=test_data['devices']['mains']['properties']['voltage'])
+    device.settings_property["param_changed"] = \
+        SettingsProperty(gateway=gateway,
+                         session=session,
+                         element_uid=f"cps.{device_uid}")
+    device.settings_property['general_device_settings'] = \
+        SettingsProperty(gateway=gateway,
+                         session=session,
+                         element_uid=f"gds.{device_uid}",
+                         events_enabled=test_data['devices']['mains']['properties']['eventsEnabled'],
+                         icon=test_data['devices']['mains']['properties']['icon'],
+                         name=test_data['devices']['mains']['properties']['itemName'],
+                         zone_id=test_data['devices']['mains']['properties']['zoneId'])
+    device.settings_property["led"] = \
+        SettingsProperty(gateway=gateway,
+                         session=session,
+                         element_uid=f"lis.{device_uid}",
+                         led_setting=test_data['devices']['mains']['properties']['led_setting'])
+    device.settings_property["protection"] = \
+        SettingsProperty(gateway=gateway,
+                         session=session,
+                         element_uid=f"ps.{device_uid}",
+                         local_switching=test_data['devices']['mains']['properties']['local_switch'],
+                         remote_switching=test_data['devices']['mains']['properties']['remote_switch'])
 
     return device