+++ docs/CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.15.0] - 2020/09/29
+
+### Added
+
+- Support for Eurotronic Spirit
+- Shared zeroconf instances can now be used
+- More units for multi level sensors
+- Prevent switching of binary switches that have active remote switching protection
+- Send changes of battery level to the subscribers
+
+### Fixed
+
+- Protection mode settings don't lead to websocket errors any more - [#65](https://github.com/2Fake/devolo_home_control_api/issues/65)
+
 ## [v0.14.0] - 2020/08/21
 
 ### Added